Add pwm.c and gpio.c for the stm32
Create pwm and gpio files (.c and .h) to basics uses. Functions only
tested on TIM1, channel 1 (PB6).
TODO1 : Test the functions on other Timer/channels
TODO2 : Prescaler and AutoReload of the Timers are set to run a 1064 Hz.
We could try to add the frequency wanted into param then calculate the
Prescaler and Auto Reload values into the function.